+++ src/main/java/cn/shuzilm/util/DeviceBlackListUtil.java
@@ -33,9 +33,9 @@ public class DeviceBlackListUtil {
             Select select = new Select();
             String date = sdf.format(new Date());
             ResultMap countRs = select.selectSingle(countSql.replace("<DATE>", date));
-            int count = 0;
+            Long count = 0L;
             if(countRs != null) {
-                count = countRs.getInteger("count");
+                count = countRs.getLong("count");
             }
             if(count > 1000000) {
                 log.error("设备黑名单数量（数量：{}）超限, 不执行操作", count);
